Code Overhaul
Split single python file into 5 segments:
Classes - contains the classes employee, shift, break, and cart duty
Parser - parses both the main schedule and the cart duty schedule
scheduler - contains the logic to schedule employees for a carty duty
writer - to write the new times to a csv to be print
main - hooks together each piece
